+++ src/main/java/com/backend/allreva/auth/application/CustomUserDetailsService.java
@@ -6,7 +6,7 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Component;
 
 import com.backend.allreva.auth.application.dto.PrincipalDetails;
-import com.backend.allreva.member.command.application.MemberRepository;
+import com.backend.allreva.member.command.domain.MemberRepository;
 import com.backend.allreva.member.command.domain.Member;
 import com.backend.allreva.member.exception.MemberNotFoundException;
