+++ src/test/java/com/backend/allreva/member/integration/OAuth2RegisterIntegrationTest.java
@@ -8,7 +8,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import com.backend.allreva.artist.command.ArtistRepository;
 import com.backend.allreva.artist.command.domain.Artist;
 import com.backend.allreva.auth.application.dto.PrincipalDetails;
-import com.backend.allreva.member.command.application.MemberRepository;
+import com.backend.allreva.member.command.domain.MemberRepository;
 import com.backend.allreva.member.command.application.dto.MemberInfoRequest;
 import com.backend.allreva.member.command.application.dto.MemberInfoRequest.MemberArtistRequest;
 import com.backend.allreva.member.command.domain.Member;
